refactor(mobile): split store files
